modifica totale

+++ src/main/java/it/contrader/dto/CityDTO.java
@@ -1,6 +1,7 @@
 package it.contrader.dto;
 
 import java.util.List;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class CityDTO {
 
-	private long idCity;
+	private Integer idCity;
 
 	private String namecity;
